Easier adds

+++ MyLists/View/Blueprints/BlueprintsListView.swift
@@ -11,37 +11,55 @@ import SwiftData
 struct BlueprintsListView: View {
     @Environment(\.modelContext) private var modelContext
     
-    @Query(sort: [SortDescriptor(\Blueprint.name)]) private var blueprints: [Blueprint]
+    @Query(sort: [SortDescriptor(\Blueprint.name, order: .forward)]) private var blueprints: [Blueprint]
     @Query private var toDoLists: [ToDoList]
-    
+    @State private var showAddBlueprintSheet = false
+    @State private var showErrorAlert = false
+
     var body: some View {
         List {
             ForEach(blueprints) { blueprint in
-                NavigationLink(destination: BlueprintItemsListView(blueprint, todoLists: toDoLists)) {
+                NavigationLink(destination: BlueprintItemsListView(blueprint)) {
                     BlueprintRowView(blueprint: blueprint)
                 }
             }
-            .onDelete(perform: deleteBlueprints)
+            .onDelete(perform: deleteBlueprint)
+        }
+        .alert(isPresented: $showErrorAlert) {
+            Alert.genericErrorAlert
         }
         .toolbar {
-            NavigationLink(destination: {
-                NewBlueprintView()
-            }, label: {
-                Button("Add blueprint", systemImage: "plus") {}
-            })
+            Image.plus
+                .foregroundStyle(Color.cyan)
+                .padding(.trailing, 4)
+                .onTapGesture {
+                    showAddBlueprintSheet = true
+                }
+        }
+        .sheet(isPresented: $showAddBlueprintSheet) {
+            NewListOrBlueprintView(isSheetPresented: $showAddBlueprintSheet, entity: .blueprint)
+                .presentationDetents([.large])
+                .presentationDragIndicator(.visible)
         }
         .navigationTitle("Blueprints")
     }
 }
 
+// MARK: - SwiftData
+
 private extension BlueprintsListView {
-    func deleteBlueprints(_ indexSet: IndexSet) {
-        for index in indexSet {
+    func deleteBlueprint(_ indexSet: IndexSet) {
+        do {
+            guard let index = indexSet.first else { throw ListError.deleteEntityIndexNotFound }
             let blueprint = blueprints[index]
             modelContext.delete(blueprint)
+            try modelContext.save()
+        } catch {
+            showErrorAlert = true
         }
     }
 }
+
 
 #Preview {
     BlueprintsListView()
